+++ LIF_gen_spike_matrix.cpp
@@ -56,7 +56,7 @@ void LIF_spike::LIF_gen_spike_matrix()
                                 {
                                         ++num_spikes[nn];
                                         spike_times[nn] = current_time;
-                                        index = floor(current_time/T_BINNING);
+                                        index = (int)floor(current_time/T_BINNING);
                                         spikes(index,nn) += 1;
                                         V[nn] = VRESET;
                                 }
